refactor: move `FillEvent` struct & update imports; update rust docs in `execution` module *
* move `FillEvent` into `execution/mod.rs` & update related imports
* update rust docs in `execution` module
