+++ src/execution/handler.rs
@@ -1,9 +1,9 @@
-use crate::execution::error::ExecutionError;
-use crate::execution::fill::{Fees, FillEvent};
-use crate::execution::FillGenerator;
-use crate::portfolio::order::OrderEvent;
 use chrono::Utc;
 use serde::{Deserialize, Serialize};
+
+use crate::execution::{Fees, FillEvent, FillGenerator};
+use crate::execution::error::ExecutionError;
+use crate::portfolio::order::OrderEvent;
 
 /// Configuration for constructing a [`SimulatedExecution`] via the new() constructor method.
 #[derive(Copy, Clone, PartialEq, PartialOrd, Debug, Default, Deserialize, Serialize)]
